+++ lib/alerts/toast_enums.dart
@@ -1,0 +1,7 @@
+/// Enum values for the Toast style types.
+enum ToastNotificationStyleType {
+  SUCCESS,
+  WARNING,
+  INFO,
+  DANGER,
+}