+++ lib/metro/constants/strings.dart
@@ -3,6 +3,7 @@ const String fileOption = 'file';
 const String helpFlag = 'help';
 const String storableFlag = 'storable';
 const String forceFlag = 'force';
+const String themeDarkFlag = 'dark';
 const String controllerFlag = 'controller';
 
 // folders
@@ -11,3 +12,5 @@ const String controllerFolder = 'lib/app/controllers';
 const String widgetFolder = 'lib/resources/widgets';
 const String pageFolder = 'lib/resources/pages';
 const String modelFolder = 'lib/app/models';
+const String themeFolder = 'lib/resources/themes';
+const String themeColorsFolder = 'lib/resources/themes/styles';