+++ lib/widgets/ny_state.dart
@@ -1,7 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:nylo_support/alerts/toast_enums.dart';
+import 'package:nylo_support/alerts/toast_notification.dart';
+import 'package:nylo_support/exceptions/validation_exception.dart';
+import 'package:nylo_support/localization/app_localization.dart';
 import 'package:nylo_support/router/models/ny_argument.dart';
 import 'package:nylo_support/router/router.dart';
+import 'package:nylo_support/validation/rules.dart';
 import 'package:page_transition/page_transition.dart';
 
 abstract class NyState<T extends StatefulWidget> extends State<T> {
@@ -29,6 +34,118 @@ abstract class NyState<T extends StatefulWidget> extends State<T> {
   /// Pop the current widget from the stack.
   pop({dynamic result}) {
     Navigator.of(context).pop(result);
+
+    this.validator(
+        rules: {"title": "required|email", "body": "required|string"},
+        data: {});
+  }
+
+  showToast(
+      {ToastNotificationStyleType style = ToastNotificationStyleType.SUCCESS,
+      required String title,
+      required String description,
+      IconData? icon,
+      Duration? duration}) {
+    showToastNotification(context,
+        style: style,
+        title: title,
+        description: description,
+        icon: icon,
+        duration: duration);
+  }
+
+  /// This validator method provides an easy way to validate data.
+  /// You can use this method like the example below:
+  /// try {
+  ///   this.validator(rules: {
+  ///     "email": "email|max:20",
+  ///     "name": "min:10"
+  ///   }, data: {
+  ///     "email": _textEditingEmailController.text,
+  ///     "name": _textEditingNameController.text
+  ///   });
+  ///
+  /// } on ValidationException catch (e) {
+  ///   print(e.validationRule.description);
+  ///   print(e.toString());
+  /// }
+  /// See more https://nylo.dev/docs/2.x/validation
+  validator(
+      {required Map<String, String> rules,
+      required Map<String, dynamic> data,
+      Map<String, dynamic> messages = const {},
+      bool showAlert = true,
+      Duration? alertDuration,
+      ToastNotificationStyleType alertStyle =
+          ToastNotificationStyleType.WARNING}) {
+    Map<String, Map<String, dynamic>> map = data.map((key, value) {
+      if (!rules.containsKey(key)) {
+        throw new Exception('Missing rule: ' + key);
+      }
+      Map<String, dynamic> tmp = {"data": value, "rule": rules[key]};
+      if (messages.containsKey(key)) {
+        tmp.addAll({"message": messages[key]});
+      }
+      return MapEntry(key, tmp);
+    });
+
+    for (int i = 0; i < map.length; i++) {
+      String attribute = map.keys.toList()[i];
+      Map<String, dynamic> info = map[attribute]!;
+
+      dynamic data = info['data'];
+
+      String rule = info['rule'];
+      List<String> rules = rule.split("|").toList();
+
+      if (rule.contains("nullable") && (data == null || data == "")) {
+        continue;
+      }
+
+      List<ValidationRule?> validationRules = [
+        EmailRule(attribute),
+        BooleanRule(attribute),
+        ContainsRule(attribute),
+        URLRule(attribute),
+        StringRule(attribute),
+        MaxRule(attribute),
+        MinRule(attribute),
+      ];
+
+      for (rule in rules) {
+        ValidationRule? validationRule =
+            validationRules.firstWhere((validationRule) {
+          if (validationRule!.signature == rule) {
+            return true;
+          }
+          if (rule.contains(":")) {
+            String firstSection = rule.split(":").first;
+            return validationRule.signature == firstSection;
+          }
+          return false;
+        }, orElse: () => null);
+        if (validationRule == null) {
+          continue;
+        }
+        bool hasFailed = validationRule.handle(info);
+        if (hasFailed == false) {
+          if (showAlert == true) {
+            validationRule.alert(context,
+                style: alertStyle, duration: alertDuration);
+          }
+          throw new ValidationException(attribute, validationRule);
+        }
+      }
+    }
+  }
+
+  /// Update the language in the application
+  changeLanguage(String language, {bool restartState = true}) async {
+    await NyLocalization.instance.setLanguage(
+      context,
+      language: language,
+      restart: restartState,
+    );
   }
 
   /// Navigate to a new route in your /routes/router.dart.