+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+
+## [2.2.0] - 2021-12-07
+
+* New validator added to NyState. Allows you to validate data in widgets.
+* Added toast notification helper
+* Refactored localization class
+* Ability to create themes and theme colors from Metro Cli
+* Pubspec.yaml dependency updates
+* Bug fixes
+
+## [2.1.0] - 2021-09-21
+
+* Fix `updateLocale` method
+* Ability to add additional Router's to Nylo
+
 ## [2.0.1] - 2021-09-18
 
 * Upgrade Dart (v2.14.0)